fix(dst): serialize fake timer rollover
Spec-first: canonical=docs/dst/design.md multi-P white-box safety and docs/dst/faults.md fake-timer remap; contract=new-epoch timer registrations survive rollover and remain enumerable for rate changes; mechanism=one epoch/list critical section with full-width timer epochs; collapse-check=faithful.

Diagnosis: fault=a timer registered after new-epoch publication could be erased by the rollover winner's later head clear; proximate=dstFakeTimersRoll CASed epoch before head.Store(nil); root=epoch and intrusive-list publication were separate synchronization domains; fixing at=timer registry state transition; wider scope justified by=n/a — structurally larger lock is the simpler singular state transition.

Invariant: kind=clause-explicit; property=every timer registered in the current epoch remains in the remap enumeration until run reset; from=docs/dst/faults.md DriftClock armed-timer contract; violation=two timers arm in the new epoch, rollover loses one, and only one deadline changes rate.

Structural-check: invariant=epoch and list head describe one registry generation; illegal-state=current-epoch timer exists outside current head chain; representable-because=epoch and head were independently atomic; decision=collapse to one short mutex-protected state transition.

Triage at 20.1: 1 match; folding dst-clock-fake-timer-roll-loses-registration, redeferring none, closing it after enforcement.

Spec: docs/dst/design.md, docs/dst/faults.md; Plan: docs/plans/dst-audit-issue-burn-down.md; divergences: none.

+++ src/runtime/time.go
@@ -90,8 +90,8 @@ type timer struct {
 	blocked      uint32 // number of goroutines blocked on timer's channel
 	rand         uint32 // randomizes order of timers at same instant; only set when isFake
 	dstHost      uint32 // DST: host id of the goroutine that armed this fake timer (set at modify under an active simulation); lets DriftClock find a host's pending timers to rate-convert on a mid-run rate change. Unused when DST off / not a fake timer
-	dstReg       uint32 // DST: run epoch in which this fake timer was registered in the per-run fake-timer list (dst.go dstFakeTimers); dedups registration and ignores a timer object reused from a prior run. Unused when DST off / not a fake timer
-	dstFakeNext  *timer // DST: intrusive link in the per-run fake-timer list (dst.go dstFakeTimers), a lock-free prepend stack; set once per run at registration. Unused when DST off / not a fake timer
+	dstReg       uint64 // DST: run epoch in which this fake timer was registered in the per-run fake-timer list (dst.go dstFakeTimers); dedups registration and ignores a timer object reused from a prior run. Unused when DST off / not a fake timer
+	dstFakeNext  *timer // DST: intrusive link in the per-run fake-timer list (dst.go dstFakeTimers); set once per run at registration. Unused when DST off / not a fake timer
 	dstWhenShift int64  // DST: base-ns the DriftClock overdue conversion moved when TOWARD now (new-old, accumulated across changes); added back to the fire-time delay so the DELIVERED timestamp keeps the original due time while the re-arm catch-up uses the converted anchor. Cleared at fire and at modify. Unused when DST off / not a fake timer
 
 	// Timer wakes up at when, and then at when+period, ... (period > 0 only)
